+++ src/main/java/edu/hm/hafner/analysis/Issue.java
@@ -14,7 +14,7 @@ import org.apache.commons.lang3.StringUtils;
 
 import edu.hm.hafner.util.Ensure;
 import edu.hm.hafner.util.TreeString;
-import edu.hm.hafner.util.TreeStringBuilder;
+import edu.hm.hafner.util.VisibleForTesting;
 import edu.umd.cs.findbugs.annotations.Nullable;
 
 /**
@@ -25,7 +25,7 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 @SuppressWarnings({"PMD.TooManyFields", "PMD.GodClass", "NoFunctionalReturnType"})
 public class Issue implements Serializable {
     private static final long serialVersionUID = 1L; // release 1.0.0
-    protected static final String UNDEFINED = "-";
+    static final String UNDEFINED = "-";
 
     /**
      * Returns the value of the property with the specified name for a given issue instance.
@@ -191,10 +191,11 @@ public class Issue implements Serializable {
      *         the other issue to copy the properties from
      */
     protected Issue(final Issue copy) {
-        this(copy.fileName, copy.getLineStart(), copy.getLineEnd(), copy.getColumnStart(), copy.getColumnEnd(),
-                copy.getLineRanges(), copy.getCategory(), copy.getType(), copy.getPackageName(), copy.getModuleName(),
-                copy.getSeverity(), copy.getMessage(), copy.getDescription(), copy.getOrigin(), copy.getReference(),
-                copy.getFingerprint(), copy.getAdditionalProperties(), copy.getId());
+        this(copy.getFileNameTreeString(), copy.getLineStart(), copy.getLineEnd(), copy.getColumnStart(),
+                copy.getColumnEnd(), copy.getLineRanges(), copy.getCategory(), copy.getType(),
+                copy.getPackageNameTreeString(), copy.getModuleName(), copy.getSeverity(), copy.getMessageTreeString(),
+                copy.getDescriptionTreeString(), copy.getOrigin(), copy.getReference(), copy.getFingerprint(),
+                copy.getAdditionalProperties(), copy.getId());
     }
 
     /**
@@ -241,9 +242,9 @@ public class Issue implements Serializable {
             final int lineStart, final int lineEnd, final int columnStart, final int columnEnd,
             @Nullable final LineRangeList lineRanges,
             @Nullable final String category, @Nullable final String type,
-            @Nullable final String packageName, @Nullable final String moduleName,
+            final TreeString packageName, @Nullable final String moduleName,
             @Nullable final Severity severity,
-            @Nullable final String message, @Nullable final String description,
+            final TreeString message, final TreeString description,
             @Nullable final String origin, @Nullable final String reference,
             @Nullable final String fingerprint, @Nullable final Serializable additionalProperties) {
         this(fileName, lineStart, lineEnd, columnStart, columnEnd, lineRanges, category, type, packageName, moduleName,
@@ -294,13 +295,12 @@ public class Issue implements Serializable {
     @SuppressWarnings("ParameterNumber")
     protected Issue(final TreeString fileName, final int lineStart, final int lineEnd, final int columnStart,
             final int columnEnd, @Nullable final LineRangeList lineRanges, @Nullable final String category,
-            @Nullable final String type, @Nullable final String packageName,
+            @Nullable final String type, final TreeString packageName,
             @Nullable final String moduleName, @Nullable final Severity severity,
-            @Nullable final String message, @Nullable final String description,
+            final TreeString message, final TreeString description,
             @Nullable final String origin, @Nullable final String reference,
             @Nullable final String fingerprint, @Nullable final Serializable additionalProperties,
             final UUID id) {
-        TreeStringBuilder builder = new TreeStringBuilder();
 
         this.fileName = fileName;
 
@@ -332,12 +332,12 @@ public class Issue implements Serializable {
         this.category = StringUtils.defaultString(category).intern();
         this.type = defaultString(type);
 
-        this.packageName = builder.intern(defaultString(packageName));
+        this.packageName = packageName;
         this.moduleName = defaultString(moduleName);
 
         this.severity = severity == null ? Severity.WARNING_NORMAL : severity;
-        this.message = builder.intern(StringUtils.stripToEmpty(message));
-        this.description = builder.intern(StringUtils.stripToEmpty(description));
+        this.message = message;
+        this.description = description;
 
         this.origin = stripToEmpty(origin);
         this.reference = stripToEmpty(reference);
@@ -423,6 +423,17 @@ public class Issue implements Serializable {
     }
 
     /**
+     * Returns the tree-string containing the name of the file that contains this issue. Typically this file name is the
+     * absolute name.
+     *
+     * @return the cached tree-string containing the name of the file that contains this issue
+     */
+    @VisibleForTesting
+    TreeString getFileNameTreeString() {
+        return fileName;
+    }
+
+    /**
      * Returns the folder that contains the affected file of this issue.
      *
      * @return the folder of the file that contains this issue
@@ -441,7 +452,7 @@ public class Issue implements Serializable {
     }
 
     /**
-     * Returns the base name of the file that contains this issue (i.e. the file name without the full path). 
+     * Returns the base name of the file that contains this issue (i.e. the file name without the full path).
      *
      * @return the base name of the file that contains this issue
      */
@@ -517,6 +528,16 @@ public class Issue implements Serializable {
     }
 
     /**
+     * Returns the tree-string containing the detailed message for this issue.
+     *
+     * @return the message
+     */
+    @VisibleForTesting
+    TreeString getMessageTreeString() {
+        return message;
+    }
+
+    /**
      * Returns an additional description for this issue. Static analysis tools might provide some additional information
      * about this issue. This description may contain valid HTML.
      *
@@ -524,6 +545,17 @@ public class Issue implements Serializable {
      */
     public String getDescription() {
         return description.toString();
+    }
+
+    /**
+     * Returns the tree-string containing an additional description for this issue. Static analysis tools might provide
+     * some additional information about this issue. This description may contain valid HTML.
+     *
+     * @return the description
+     */
+    @VisibleForTesting
+    TreeString getDescriptionTreeString() {
+        return description;
     }
 
     /**
@@ -580,6 +612,17 @@ public class Issue implements Serializable {
      */
     public String getPackageName() {
         return packageName.toString();
+    }
+
+    /**
+     * Returns the tree-string containing the name of the package or name space (or similar concept) that contains this
+     * issue.
+     *
+     * @return the package name
+     */
+    @VisibleForTesting
+    TreeString getPackageNameTreeString() {
+        return packageName;
     }
 
     /**