+++ src/main/java/edu/hm/hafner/analysis/IssueBuilder.java
@@ -9,6 +9,7 @@ import org.apache.commons.lang3.StringUtils;
 import edu.hm.hafner.util.PathUtil;
 import edu.hm.hafner.util.TreeString;
 import edu.hm.hafner.util.TreeStringBuilder;
+import edu.hm.hafner.util.VisibleForTesting;
 import edu.umd.cs.findbugs.annotations.Nullable;
 
 import static edu.hm.hafner.util.IntegerParser.*;
@@ -30,9 +31,7 @@ import static edu.hm.hafner.util.IntegerParser.*;
  */
 @SuppressWarnings({"InstanceVariableMayNotBeInitialized", "JavaDocMethod", "PMD.TooManyFields"})
 public class IssueBuilder {
-
-    private static TreeStringBuilder treeStringBuilder = new TreeStringBuilder();
-
+    private TreeStringBuilder treeStringBuilder = new TreeStringBuilder();
     private int lineStart = 0;
     private int lineEnd = 0;
     private int columnStart = 0;
@@ -229,15 +228,27 @@ public class IssueBuilder {
      * @return the created issue
      */
     public Issue build() {
-        Issue issue = new Issue(treeStringOfFileName(fileName), lineStart, lineEnd, columnStart, columnEnd, lineRanges, category, type,
-                packageName, moduleName, severity, message, description, origin, reference, fingerprint,
+        Issue issue = new Issue(treeStringOfFileName(fileName), lineStart, lineEnd, columnStart, columnEnd, lineRanges,
+                category, type, treeStringOfPackageName(packageName), moduleName, severity,
+                stripToEmptyTreeString(message), stripToEmptyTreeString(description), origin, reference, fingerprint,
                 additionalProperties, id);
         id = UUID.randomUUID(); // make sure that multiple invocations will create different IDs
         return issue;
     }
 
-    protected static TreeString treeStringOfFileName(@Nullable final String fileName) {
-        return treeStringBuilder.intern(normalizeFileName(fileName));
+    @VisibleForTesting
+    TreeString treeStringOfFileName(@Nullable final String str) {
+        return treeStringBuilder.intern(normalizeFileName(str));
+    }
+
+    @VisibleForTesting
+    TreeString stripToEmptyTreeString(final String string) {
+        return treeStringBuilder.intern(StringUtils.stripToEmpty(string));
+    }
+
+    @VisibleForTesting
+    TreeString treeStringOfPackageName(@Nullable final String str) {
+        return treeStringBuilder.intern(defaultString(str));
     }
 
     private static String normalizeFileName(@Nullable final String platformFileName) {