+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Added a parser for CMake warnings.
 Added a parser for JSON output from Cargo.
 
 ### Fixed
-- [JENKINS-56333](https://issues.jenkins-ci.org/browse/JENKINS-56333): 
+- [JENKINS-56333](https://issues.jenkins-ci.org/browse/JENKINS-56333),
+[PR#129](https://github.com/jenkinsci/analysis-model/pull/129): 
 MsBuild Parser: Treat errors as errors and not warning (high).
 - [JENKINS-56737](https://issues.jenkins-ci.org/browse/JENKINS-56737),
 [PR#136](https://github.com/jenkinsci/analysis-model/pull/136)
@@ -21,6 +22,12 @@ Javac parser: Add the ability to parse warnings with preceding timestamps.
 - [JENKINS-56214](https://issues.jenkins-ci.org/browse/JENKINS-56214),
 [PR#142](https://github.com/jenkinsci/analysis-model/pull/142)
 ChackStyle parser: Map errors to severity ERROR (rather than WARNING_HIGH)
+- [JENKINS-48647](https://issues.jenkins-ci.org/browse/JENKINS-48647),
+[PR#145](https://github.com/jenkinsci/analysis-model/pull/145): 
+MsBuild Parser: Update regular expression to detect logging prefixes.
+- [JENKINS-52477](https://issues.jenkins-ci.org/browse/JENKINS-52477),
+[PR#146](https://github.com/jenkinsci/analysis-model/pull/146):
+FileReaderFactory: Detect charset from XML-header when not specified. 
 
 ### Changed
 - Filters now work on a substring of the property, you don't need to create a regular